Some warnings removed

+++ src/tkgate/and.c
@@ -343,7 +343,7 @@ int AOX_GateParmList(FILE *f,GCElement *g)
 	  fprintf(f,"%s",w->net->n_signame);
 	fprintf(f,")");
       } else
-	fprintf(f,"(%sw%x)",inv,(unsigned)w->net);
+	fprintf(f,"(%sw%x)",inv,(intptr_t)w->net);
     }
   }
   fprintf(f,");");
